Implement non-blocking FUSE_INIT request
In the original macfuse mount() syscall is blocking.
It blocks until a user daemon responses with FUSE_INIT.
This does not match the upstream fuse behavior as well as
requires a few tricks (e.g. double-fork for fuse_mount)
in the userspace library.

The current change makes FUSE_INIT asyncronious.

There is a thing to remember: right after vfs_mount function ends but
before kernel resurns from mount() the kernel requests FUSE_STATFS. So
in case if fuse daemon is not connected (and the session is not initialized)
yet we need to fake STATFS.