Fix deadlock when biglock is enabled (64bits kernel).
What happens here is when EncFS processes FUSE_CREATE requests it opens a file
in ".crypt" directory. Although the file is not under fuse4x it causes
VNOP_FSYNC for fuse4x files. It sounds like XNU has some postponed fsync mechanism.

To avoid this problem as well as reduce biglock contention we release it before process
goes to sleep (and waits for a response from fuse daemon).

Here is a repro case for EncFS deadlock issue:

int main() {
  char filename[300];
  for(int dir = 0; dir < 100; dir++) {
    sprintf(filename, ENCFS_DIR "dir-%i", dir);
    mkdir(filename, 0777);

    if (chdir(filename) < 0)
      perror("chdir");

    for(int file = 0; file < 10000; file++) {
      printf("creating dir-%d/file-%d\n", dir, file);

      sprintf(filename, "file-%i", file);
      int fd = open(filename, O_CREAT, 0660);
      if (fd < 0)
        return EXIT_SUCCESS; // we successeded in killing encfs, yay!
      close(fd);
    }
  }

  printf("There is no deadlock!\n");
  return EXIT_SUCCESS;
}