Added SYMROOT=build SHARED_PRECOMPS_DIR=build -PBXBuildsContinueAfterErrors=0 to the xcodebuild call
Because build.rb assumes the results to live below "build" it's necessary to tell xcodebuild that SYMROOT is really
there. If you have Xcode configured to have a shared build directory for all projects (like I do), then this is
not the case and setting SYMROOT accordingly is required. SHARED_PRECOMPS_DIR is merely an optimization and not
strictly required. Also, stopping the build on first error isn't required also (but comes in handy IMO).