Compile final distribution kext with 10.5/10.6 SDK
Unlike user-space binaries MACOSX_DEPLOYMENT_TARGET does not work for kext [1].
Kext provided for 10.5 have to be compiled against SDK 10.5. So we introduced a
new configuration that hardcodes dependency to this SDK. We don't want to hardcode it for 'Release'
xcode configuration because we want to leave it SDK-independent (for Homebrew/macports/developers).

Actually we have to use 10.5 only for 32bit platform. 64bit kext requires SDK 10.6, so we provide
different build-time SDKs for different architectures.

This fixes https://github.com/fuse4x/fuse4x/issues/15

Introduce --release build flag.

[1] http://web.archiveorange.com/archive/v/8tBh4VFfPneDQmDJNnl6