Add a stub for VNOP_ALLOCATE
On vnop_allocate (file space preallocation) do nothing and just return success.
This is needed for programs that preallocate file space ahead.
If fcntl with PREALLOCATE is called then fuse4x returns an error,
and a user program thinks that media does not have enough space. But
in fact fuse4x returns ENOTSUP (syscall is not supported).

Add a stub that returns success and makes the user-space programm believe that
media has enough space.

The long-term solution is to add FUSE_FALLOCATE fuse request type and implement it
in kext.